Packages: move SequentialCellGenerator to .clip and CutoutTextCell to .cell
