+++ indicator/src/main/java/com/fuzz/indicator/CutoutViewIndicator.java
@@ -42,7 +42,7 @@ import com.fuzz.indicator.proxy.ProxyReference;
 import com.fuzz.indicator.proxy.StateProxy;
 import com.fuzz.indicator.proxy.UnavailableProxy;
 import com.fuzz.indicator.proxy.ViewProvidingStateProxy;
-import com.fuzz.indicator.text.CutoutTextCell;
+import com.fuzz.indicator.cell.CutoutTextCell;
 
 import static android.view.ViewGroup.LayoutParams.WRAP_CONTENT;
