+++ indicator/src/main/java/com/fuzz/indicator/text/TextCellGenerator.java
@@ -27,6 +27,7 @@ import com.fuzz.indicator.CutoutViewIndicator;
 import com.fuzz.indicator.CutoutViewLayoutParams;
 import com.fuzz.indicator.cell.CutoutCell;
 import com.fuzz.indicator.CutoutCellGenerator;
+import com.fuzz.indicator.cell.CutoutTextCell;
 
 /**
  * This generator is intended to animate stuff on TextViews
